Add files via upload
Initial analysis for the chip I purchased from ICStation reveals that the the Output of the chip I received averages 3.4 volts when motion is detected and stays high for approximately 2.98 seconds.   Because of the inefficiency of the code written, I was only able to collect 2510 samples when the RCWL-0516 is high.